Add files via upload
Hi everyone, my name is Zakiy Nashrudin Wahid. So, this is the first project that I ever posted on my GitHub account. In this project, I made my personal portfolio website, and this is the first time I have built my project completely from scratch since I entered the programming world three months ago, even though I looked into tutorials on YouTube to program this project. I completed this project in 5 days by coding the HTML for the contents of the website, styling the website with CSS to make it look beautiful and completely responsive for all devices, and then making the Javascript code to make it interactive with the user. I also use Google Spreadsheet to store the data form from the web by linking it to Javascript code. 

So for the summary, the tech stack that I have used is:
1. HTML for the website content
2. CSS for styling the website
3. Javascript for making interactive functions with the user
4. Google Spreadsheet to store my data-form from my website
Although the project has been considered done, it can still edited sometimes for improvement. I will update the web design, add more features to make it more interactive and engaging for the users, and post another project to fill the content of my portfolio web also haha.

Note: all the experience and the educational background are just fiction.